Initial update

+++ app/src/main/java/table_entities/characterLanguages.kt
@@ -28,6 +28,6 @@ import androidx.room.PrimaryKey
 data class CharacterLanguage(
     @PrimaryKey(autoGenerate = true)
     val characterLanguageId: Int = 0,
-    val characterId: Int,
+    val characterId: Long,
     val languageId: Int
 )
